requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
sqlalchemy 
databases

langchain
langchain_core
python-dotenv
langchain_community
langchain-huggingface

langserve
sse_starlette

pypdf
faiss-cpu